fix(chats): split chat-reactions runtime into its own extension
ConnectionRuntime grew past SwiftLint's type_body_length threshold
(2500 lines). Move the 3.2 chat-reactions helpers to
ConnectionRuntime+ChatReactions.swift so the main type stays under
the limit and the new surface lives in a dedicated file alongside the
existing runtime extensions.
